access: let existing phabricator-root-admins get on aphlict machine
The existing phabricator admins with root access should also have
access to the newly created machine for aphlict.

A new role has been created which is a subset of the existing role.

So this is not new access but just updating it to the new role name.

The non-root phabricator admins and the bulk job manager don't need
this i think, so they are skipped for now.

Bug: T238593
Change-Id: Icf7fc2da8c6f397254e64ae1b59a921670189268