nutcracker: drop puppet < 3.5 support
currently the nutcracker module tries to check a fact which no longer
exists to determine if there is support for puppet <= 3.5.  We no longer
support puppet < 5.5 as such we can drop this check

Also add types and drop deprecated validate_ functions
the Verbosity test is Varieant[String, Integer] to make this first
change a noop.  See follow up CR's where this is dropped

Hosts: C:nutcracker

Bug: T258931
Change-Id: Ia7b8a8f26cc069859c2a5d6fc23b0b37d23e9be2