nutcracker: ensure verbosity is an integer
The nutcracker::verbosity parameter has been set as the string '4'.
however this parameter should be an integer.  The default value is also 4
and therefore dose not need to be set explicitly.  Further the value is
set in hiera using the nutcracker module namespace and not a profile::
namespace

Hosts: C:nutcracker

Bug: T258931
Bug: T247956
Change-Id: I26c4d466d841de793a542c556887ce771df5d28d