dynamicproxy: Add custom response for missing proxy backend
In urlproxy mode we are using a wildcard DNS setup for *.wmflabs.org
that will respond with a 404 when we do not have a configured backend
for the current hostname. This change adds a new page using the
mediawiki::errorpage define to be served for that "missing proxy" case
rather than the boring default nginx 404 page.

Bug: T258730
Change-Id: I9a28dbbc5f8157db4358c80a14c0bdbc9551b354