admins: add all members of wdqs-admins to wdqs-roots
Simply copy all members of the existing wdqs-admins group
and add them to the wdqs-roots group to give them full root.

This is the first and most simple step to move forward on
T258739 which will then need some more follow-ups to update
and clean references to the admins group.

As mentioned in comments on I181b230db3a2 which should be rebased
on this.

Bug: T258739
Change-Id: Ifc68e3a1f43521b128a9c2d29b8f89dc20be8bea