airflow: Reduce scheduler execution rate to one minute
The scheduler was runing each dag every 5 seconds to see if something
new needs to be scheduled. We really don't need that level of
granularity, and it spams up the logs with execution information 12
times a minute. Reduce the execution frequency to per-minute which
should be more than sufficient for our needs.

Change-Id: I460bdb460a7f515ac0c3595e4df7fb01217d9f13