sudo: create new command to create safe wildcard sudo commands
sudo commands using a wildcard in the path expantion are suceptible to
wild card expantion e.g. with the following:
  'ALL = NOPASSWD: /bin/rmdir -p /srv/*'
A user is able run the following
  rmdir /srv/../
  rmdir /srv/foo /

This command takes the command and path argument and returns a command
spec which is a bit more resistent to this misconfiguration.  taking the
above example the function :
  '/bin/rmdir'.sudo::safe_wildcard_path('/srv/*')
The following string is returned
  /bin/rmdir /tmp/*, !/bin/rmdir /tmp/* *, !/bin/rmdir /tmp/*..*

Bug: T258943
Change-Id: Ifec9c02df1c511797fd2d20e9683c579451dbc30