mariadb: Match port 3351 and 3352 to 2 analytics sections
matomo and analytics_meta used to use already-assigned ports for
exposing to mysql. This was an issue, because it meant that just
from the port, it is not possible to know if they are misc "m1"
and "m2" instances or analytics instances.

Now that they have been moved to 3351 and 3352 respectively,
match them on monitoring.

Bug: T234826
Change-Id: I4422c566da1641b00589316cd6660e9c643718e4