mariadb: Reduce buffer cache memory footprint to prevent OOMs
dbstore1004 showed clues of starting to swapp/heavy memory usage
due to some ooms. Reduce the amount of memory in use for some
of the processes to avoid process failures.

Reorganize so s4 has the most memory, as it was the one demanding
it and traditionally will require more for both replication writes
and heavy querying.

Change-Id: I3eb27726cc38fda6a92551056ccf5eedb6d7f70c