Set spark deploy-mode client for all the Analytics Hive to Druid jobs
The current deploy-mode cluster is very handy since it avoids to run
the Spark driver process on the client (having it running on hadoop instead),
but the drawback is that there is no clear error propagation when
a job fails. In fact when a job fails in hadoop, the correspondent timer
succeeds since there is no propagation of the failure.
This change forces the driver to run locally on an-launcher1002, having
a proper return code (and hence timer unit errors, icinga errors, etc..)
when an indexation fails.

Bug: T254493
Change-Id: I9563a26b394cd8d006930c9614693d076848ecd0