mediawiki::maintenance: install modsecurity-crs
A follow-up to I10f18f2f4e539b which installed
libapache2-mod-security which was a follow-up
to I6a682e14384fba to add the server-header config
on mwmaint servers just like they are on mw servers.

If libapache2-mod-security is installed without
also installing modsecurity-csr there is a
syntax error and httpd dies on the next restart.

The reason is a bug in the former package which
installs /etc/apache2/mods-available/security2.conf
with the lines:

  # Include OWASP ModSecurity CRS rules if installed
  IncludeOptional /usr/share/modsecurity-crs/owasp-crs.load

Clearly the intention is here to make this optional and only include
it if it exists.

The bug is that this syntax does not actually make it optional
in httpd config.

Upstream httpd docs say:

" It works identically to the Include directive,
but it will be silently ignored (instead of causing an error)
if wildcards are used and they do not match any file or
directory or if a file path does not exist on the file system."

So because this does _not use wildcards_ it means IncludeOptional
behaves like an Include despite its name. Yay!

This then leads to:

file /usr/share/modsecurity-crs/owasp-crs.load: No such file or directory

which then becomes:

Syntax error on line 12 of /etc/apache2/mods-enabled/security2.conf

This does not seem obvious at all. You would think if you wanted
non-optional you would just use regular Include, but #httpd on Freenode
said there are historic reasons for this.

https://httpd.apache.org/docs/2.4/mod/core.html#includeoptional

Bug: T255629
Change-Id: I174b9a7b2ae9789dd77de35b6094a7e95af82c77