haproxy-galera: Make a meaningful healthcheck
A tcp healthcheck does not determine if the DB is at all usable. This leads
to a lot of problems when any database server needs to rebuild or needs to
be restarted.

This introduces a simple http check that uses prometheus db credentials
to connect and check if the DB is read_only and if it is wsrep_ready.

If it is not read-write and not ready it will not send connections to it.
Also, an admin can create a file in /tmp that will cause it to start throwing
404s, and thus depool the node.

This is effectively using systemd as a xinetd replacement, tested locally.
Before merge we should copy the script over and ensure it returns as expected
when run as prometheus user.

Also: remove the sticktable config because that might be causing the
connections to always end up on only one server.

Change-Id: Ia393fa1bc5561db48e9696c9094d57e40719d8c4