cumin: Update wmf_auto_reimage_lib for the new httpbb test layout
Now that httpbb-tests is broken up by host type, update the --httpbb
flag to run the tests under httpbb-tests/appserver/.

This is the minimal change for backwards compatibility, since appserver
tests used to be all we had. In the future we might update this to
specify the tests to run, like "--httpbb=appserver", in order to make
the other tests available.

Bug: T259665
Change-Id: Icd380be5dddb9d8a066d595cb9b1371c1b657e4b