install: Prevent full wipe of dbprov2003 data by changing its recipe
dbprov2003 has been fully installed with the db.cfg recipe.
Now change it to the "reuse" one, which should prevent losing the
/srv partition by accident.

Keep the dbprov1003 (not yet racked) with the full wipe one until
it is installed.

This will be tested subsequently with a reimage now that there is
no significan data that could be lost.

Bug: T257551
Bug: T252027
Change-Id: I2ef9ce207eb8f4c039102b42ff9ab31ac0651bf3