shared-storage: remove nfs-manage-binds
Since a big blocker on failover has been bind mounts, we got rid of them
since the only reason to practice that sort of pattern in NFSv4 is
for discoverability anyway (which is unnecessary when all mounts are in
puppet). Removing the script keeps admins from following old docs and
running it by mistake.

Bug: T169570
Change-Id: I3f4a5e27f25c1d8a8fe9f63ff902d4a7c714c78a