ATS: set caching to 'websockets' for Phabricator
follow-up to I472913a95d

ws://phabricator.wikimedia.org and https://phabricator.wikimedia.org
are both supposed to work but with different replacements.

What we currently see is:

phab->aphlict:22281 is working but public->wss://phabricator.wikimedia.org:443
isn't upgrading it websocket it's just treating it as a normal https request.

Comparing the config with Etherpad as another service using wss:// I noticed
this setting wasn't done for Phabricator.

So this is hopefully correct to set even if we have both https:// and wss://
connections to the same host name.

Bug: T238593
Change-Id: I3637733bd65d42c7064e710ec013f4fcdfa01275