Gemfile: match Buster puppet version
Debian Buster ships with Puppet 5.5.10, the loose constraint would bring
in a more recent version of Puppet which might unexectedly fail. That
happened with 5.5.21 which broke rspec-puppet 2.6.

Since CI uses a snapshot of bundle install, whenever the CI Docker image
is rebuild an arbitrary version of Puppet is bundled in the image and is
not updated on a per build basis.  Hence, pin to the same version
provided by Debian Buster.

Bug: T260063
Change-Id: I767b512a6bb8d55cfd53d54a52892faafc182f8b