arclamp: send mail from generate-svg cronjob to perf-team
Every once in a while we still get emails to root@ with:

ERROR: No stack counts found
ERROR: No stack counts found

As pointed out by Dave Pifke these point to an actual
problem and we should not discard these errors like
we discard stdout.

But instead of mailing all of SRE at root@ we should
forward them to the perf-team so they can react earlier.

The puppet cron type does not directly have a parameter
for mail_to but it has "environment" property which we
should be able to use likt this to achieve that.

Change-Id: I4272c70473c5f649fa86fb1639ba1b709b20df6c