releases: allow http connections also from cumin masters
httpbb is installed on deployment servers and cumin masters
but the firewall on releases only allowed connections from deployment
servers leading to failing tests when running httpbb checks from cumin.

Add cumin masters to the firewall hole and also simplify the syntax
of the srange for existing deployment servers. Separately listing
IPv4 and IPv6 with resolve should hopefully not be needed anymore.

Bug: T247652
Change-Id: I5b72a088af1d7fe13ee794485152692fe4fc8d48