aphlict: listen on IPv6 instead IPv4 for client and admin ports
The aphlict service is listening only on IPv4 but we have IPv6
DNS records so clients are trying to connect there first.

Tried just skipping the "listen" parameter entirely because
aphlict docs call it optional but when doing that the service
is not restarting.

Just using IPv6-only seems to work though.

Doing it for both the client port and the admin port where
phabricator connects to aphlict.

Bug: T238593
Change-Id: I621bc4cce27b71c6c2f6cc5e8b4bc327cecf20eb