mariadb-backups: Remove replication lag avg check from backup sources
Backup sources have its replication frequently stopped in order to
speed up backup taking.

Average replication check in the last 5 minutes could be a useful
check with a larger parameter (e.g. 1-2 hours of lag), but because
prometheus metrics used employ show slave status, which returns null
while replication is stopped (unlike the prefered pt-heartbeat,
which still works when replication is stopped), this check is not
useful for most of the time on these servers.

Note this also affects dbstore hosts that are not backups, but I think
a similar rationale applies.

Remove the check to avoid having constant UNKNOWN state checks. We could
reenable it at a later time when we have pt-heartbeat metrics available on
prometheus.

Bug: T253120
Change-Id: I11a8c3372d79756e9e1060bf434b626aba4f9bc2