Add yarn.nodemanager.vmem-pmem-ratio setting to Hadoop test
While testing spark2 running in client mode, the Yarn AM container
was often killed due to excessive virtual memory usage (but not RSS).
This behavior came out with Hadoop 2.8.5, so the speculation (not proven)
for the moment is that more/bigger shared libraries are causing more
virtual memory usage. The yarn.nodemanager.vmem-pmem-ratio's default
is 2.1 (2 to 1), this change increases it to 5.1 for hadoop test.

Bug: T244499
Change-Id: I8d4115e3fc2b7e2cfbe218442ba6f307d4823d4b