ATS: temp. set backend for releases-jenkins to releases1001
Temporarily hardcode releases-jenkins to backend releases1001
instead of the discovery record.

This allows for moving releases.wikimedia.org to a new server
without also having to configure a new jenkins at the same time.

The TLS certificate used on envoy also contains this host name
so it should be no difference.

Bug: T247652
Change-Id: Ic70dca08a3dd7305e54ee96178de11231856e40f