jenkins: redesign the way we manage service ensure/enable
We want to make sure that jenkins is never automatically started
on a new instance without explicitly enabling it.

We also want to make sure we can both stop and mask it.

At any given time it should only run on a single server and not
any others and it is used on both contint* and releases* servers.

Therefore:

- add the ability to mask a service
- change the default to be stopped and masked
- stop using 'unmanaged' which isn't an upstream parameter
  for a service
- allow exactly the values that are allowed upstream
  and pass them through
- simplify code and avoid inventing our own values
  and having additional logic
- explicitly start the service in Hiera on a host name
  basis for contint2001 and releases1001 which are currently
  the active servers
- fix mediawiki::releases profile to pass on the jenkins
  related parameters while only doing lookups from the profile namespace

A follow-up change could do this based on a global "active_server"
setting in Hiera which is compared to FQDN in puppet code, if desired.

Change-Id: Iae622499aa70e066b32ed60df93f2618eca57cfc