releases: rsync needs to be on all servers incl the primary
The rsync::quickdatacopy code needs to be applied on ALL
releases servers, not only all secondary servers.

Inside it there is code to compare source and dest server
name against FQDN and change behavior based on that.

We need it to setup the rsyncd source config to be able to pull
from the primary server.

Bug: T247652
Change-Id: I576032a16912603fcee5e031fed4073a3be91a5c