domainproxy: enforce TLS by default
* New profile::wmcs::novaproxy::do_https hiera setting
* Default profile::wmcs::novaproxy::do_https to true
* Pass profile::wmcs::novaproxy::do_https to ::dynamicproxy as its
  https_upgrade parameter.

The nginx config template for the domainproxy mode of dynamicproxy
already contains the needed stanzas to do "safe" redirection from http
-> https for GET and HEAD requests. This leaves the notorious "POST
loophole" open rather than introducing the POST-safe 307/308 redirect
codes because there are still a lot of user-agents in the wild which
don't deal well with the 307/308 codes (:sad_panda:).

The existing nginx config will also set an HSTS header with
max-age:86400 (1 day). This header will be sent for both HTTP and HTTPS
connections because of limitations in the handling of `if` blocks in
nginx. HSTS headers being included HTTP responses without TLS is
a violation of RFC 6797 section 7.2 (T102367#5056920), but there is not
a great fix for this until we can close the POST loophole.

Bug: T120486
Change-Id: Ic1ddce013bd2f743c4271fdd57fa7cf2313cf0eb