mariadb-backups: Enable eqiad backups, which are sent to codfw (backup2001)
So far, backups were taken in the mostly passive datacenter (dbprov2XXX
hosts) and sent cross-dc to eqiad (backup1001).

Now configure the opposite method: take backups generated on eqiad
(dbprov1XXX hosts) and store them long term on backup2001.

This will increase the redundancy while not affecting that much
the availability, as the backups generated on each dc should be
equivalent (something discussed widely with DBAs and other SREs),
as well as the possible impact on network bandwidth usage.

Additionally, make sure if in the future dbprovs or backup hosts
are setup in other places than eqiad or codfw, they fail very
verbosely with a parse error.

Bug: T79922
Change-Id: I4e21d8dc9356520167cf1e7280ddd1dc2b027a4d