mariadb: Setup section->port assignment on puppet
section->port assignment (and viceversa) is needed in many places
both in puppet and automation. Add it to hiera so it is easily
accessible from both.

To have a first example application, apply it to cumin hosts so
a file '/etc/mysql/section_ports.csv' exists with such information,
so it can be read by the wmfmariadbpy package.

It will later be able to be applied to other profiles, as well as
be used by several puppet classes.

Bug: T257033
Change-Id: I6d39508f7fde6406b11d5e34945ee8d36bb58dd1