wikidough: increase TCP connection limits for dnsrecursor
Set max_tcp_per_client that defines the maximum number of TCP
connections allowed from the client to 0 (pdns-recursor's default is 0
as well; 0 means unlimited) as we already limit connections from
localhost. The default value in the dnsrecursor module is 100.

Also increase the client TCP timeout to 5, that signifies the time to
wait for data from TCP clients for pdns-recursor. The default and
current value is 2; the updated value of 5 is to experiment if
increasing it decreases the timeouts observed from the dnsdist frontend.

Bug: T252132
Change-Id: I2b3ac641a007a8eaac8cbb3a469d08b1e2af6c3d