wmcs: collect prometheus metrics from alertmanager in metricsinfra
Monitoring the alertmanager seems like a good thing in general. It is
also a requirement of setting up the 'karma' front-end for alertmanager.
Karma expects to be able to determine the currently running version of
alertmanager by asking prometheus for an "alertmanager_build_info"
metric.

Patch also fixes usage of ::prometheus::server for refactoring done in
5c67ed9.

Change-Id: I5eb88a878cee0784da1cfdf3761a233e28946050