phabricator: Add new db user for the daemon, separate from web request user
Add a new user, phabricatorphd, with equivalent grants to the app user,
but that will be used for access that is not 1:1 to a webrequest.
This could be run in the future from a separate host, or have different
grants.

Also fix some discrepancy on grant differences betwen proxies for phuser.

Bug: T146055
Bug: T250361
Change-Id: I7ed8244c22bfd57ab38d113b3902d9d52dd3c73d