cloudvirt1024: move to new role, 'virt_ceph_and_backy'
This is a ceph-hosted cloudvirt, but it also runs backup jobs
to store ceph backups on the (now idle) ssds that previously
hosted VMs.

This host has a bit of hand-tuning to move the big empty partition
from /var/lib/nova/instances to /srv/backy2 and to increase the size.

Bug: T260692
Change-Id: I01fb76846bade231b3b7d7fe523cc57e9af0407c