profile::service_proxy::envoy: inject XFP in all calls to mediawiki
Since we're now using forceHttps=yes, not injecting the
X-Forwarded-Proto header would mean we get a redirect to the public
URL when calling internally and XFP is not present in the application
request.

While this is suboptimal, it worked for restbase because RB can just
go to the public URL, while it doesn't work for termbox/other services
in k8s since they're firewalled.

Change-Id: Idfb818d3a55c37d91286d6d8997806cd39968760