services_proxy::envoy: completely overwrite X-Forwarded-Proto if needed
This is a proxy for internal requests, and the requests are properly
encrypted in-flight. Envoy tries to DTRT and if no x-forwarded-proto is
provided by the request, it injects XFP: http for all requests coming
to an http endpoint (like the ones to the proxy).

So if you do a request like follows:

 $ curl -H 'Host: en.wikipedia.org' http://localhost:6500

it would result in (without xfp injected):

x-forwarded-proto: http

if we just inject the xfp: https header, it would be:

x-forwarded-proto: http,https

if we add the removal of the request header, we get

x-forwarded-proto: https

On the other hand MediaWiki relies on XFP to redirect users to https,
which of course we don't want. So in addition to add XFP: https, we
also want to remove the original XFP, as MediaWiki searches for an
exact match.

Change-Id: I055834c4291b720d6575d18f65ddafc9c8df7069