varnish: add wikilovesmonuments.org to domains allowed to access maps servers
The intention here is to allow maps.wikilovesmonuments.org to use
maps.wikimedia.org. So adding it to the regex of referrers.

Note that the domain is owned by WMNL and not WMF.
I am not sure if that matters here.

Of course we want to support 'Wiki Loves Monuments'.

It is somewhat high priority because the photo competition starts
in September and currently the site gets blocked from showing maps.

Bug: T260520
Change-Id: Ia25bcc9ed346b794174d292d7fce8513e4610a99