cumin: for new wmcs. prefix for cookbooks, grant access to wmcs-admins
In order to manage an expanded and more complex wikireplicas cluster,
people on the clinic-duty rotation for WMCS who do not have the ops group
could really use access to run the new cookbooks we are working on adding
to spicerack.

To ensure that only the default configuration is used, use a small
wrapper script to restrict the use of --config-file arguments.

Bug: T260389
Bug: T261145
Change-Id: Ifba61616be8d0976b0f48818fe1b76129aeb68ad