wikidough: enable DNSSEC validation in pdns-recursor
Based on the discussions in task T259816, enable DNSSEC validation in
Wikidough. See the task for the details but to summarize: we set
pdns-recursor to perform strict validation, which means that we validate
DNSSEC records irrespective of whether a client (like Firefox) asked for
it and we send SERVFAIL on invalid/bogus responses.

This behaviour is also consistent with that of other public resolvers.

Bug: T259816
Change-Id: Ic116f305b5ba7a39025316a7ca32ee788b3ef149