testreduce: add parameters to control client services
In addition to vd_server and rt_server we also need parameters
to control the status of the clients, vd_client and rt_client.

Despite them being clients they are also systemd services
that have a running/stopped state and if they run while the
servers are not then it spams logs with errors.

Bug: T257906
Change-Id: Ibd3bce7ead49c48654f8fd1993a57363e6e4cbec