doc: set  WMF_DOC_PATH environment variable
CI publishes the generated documentation inside the document root and
will be moved outside (to /srv/doc). We need a way to pass the
destination to our PHP boilerplate and an environment variable seems
like a good way for it.

SetEnv would set it AFTER the rewrite rules and we will need to reuse
that variable in some rules. Hence uses SetEnvIf instead which would
apply on any request URI.

Bug: T149924
Change-Id: If98165f512ac505614e7b18e43b91a6db9431376