prometheus: hiera() -> lookup(), add data type for prometheus_nodes
Replace hiera() with lookup() in general because it will be
deprecated in future puppet versions.

Also add data types to parameters which don't have them.

Also a general thing, just picking prometheus-related classes
in this patch because they are all similar and to keep the patch
a reasonable size, not too small and not too large.

There will be some follow-up patches to add more data types and similar
small improvements.

Change-Id: I3fc8b717908cbd940e209930867631957e6d9a3f