openstack: replace hiera() with lookup() in several places
Part of a general cleanup to replace hiera() with lookup()
across the puppet repo because it will be deprecated in future
puppet versions.

Since none of these parameters have default values (which is
a good thing!) this should
be the low-hanging fruit where you really just have to replace
hiera with lookup and it's a complete noop.

Change-Id: I5d95f1e4532a44bed33dbcd2b534e32e8839eb1a