hiera: change mwdebug1001 opcache settings for testing
We are trying to reproduce the opcache corruptions
we are experiencing from time to time. In order
to do that we need to fiddle with the opcache settings
a bit, and remove the php-fpm restart timer we have
as a counter measure. We are using mwdebug1001 as
the target host.

Bug: T261009
Change-Id: If93b0f3d7d09d04591117daf327416a5eef99c55