dynamicproxy: allow service workers in Toolforge
Follow up to 2cb73aa. When we were serving all Toolforge tools from the
same hostname (tools.wmflabs.org) it was very dangerous to allow a tool
to register a service worker outside of the tool's subdirectory scope by
sending a Service-Worker-Allowed header. Now that each tool has
a separate hostname we can remove that restriction.

Bug: T158216
Change-Id: I286ae60f1330b11f5d2c58a4c7f406ea89250c4a