Retire stub firejail code in service::uwsgi
I was wondering why firejail was installed on debmonitor when deploying
the latest firejail security update. Remove the require_package and
the unused parameter, this was added in April 2016 and these days we
have better means to add hardening to an UWSGI service via the systemd
service unit's hardening options.

Change-Id: Ib52d880cfe9be4be566320c73da8d1384f3257b4