wikireplicas: create multiinstance roles and profiles
This should establish web and analytics multiinstance setups for the
replicas. This is just the install framework
for the hosts and is leaving out the complications needed to configure
the proxies and the admin scripts used by WMCS to be done in future
patches.

Adjusted a bit for the work done on the core multiinstance profile so far.
This is also dependent on I175cfc93f652e9c0ec12e, however this patch
doesn't directly assign the new roles to servers, so merge should be possible.

This also is moving the scriptconfig bit out of the roles and into the
profiles with require statements. Also trying removing the passwords
module pending the results of If35f54760448b3ed

Bug: T260843
Change-Id: If6de864968db6273f894a00bd2ef9ed929dc86a1