ripeatlas: bump IPv6 failure threshold
Looking at the past few days of spam, this ought to be enough to avoid
spam but still catch widespread issues.  The baseline success rate for
IPv6 is closer to 92% anyway (more like 99% for IPv4).  This moves the
reporting threshold to something like 10-11% depending on site (as # of
probes varies by anchor).

Change-Id: If9fcd8eeeb785d1694f26e09a07806979b414a7a