shared-storage: add specific NFS volume monitoring for cleanups
While we could change the parameters around the disk check for the servers,
that doesn't correctly recognize that we need different monitors for the
shared volumes. The thresholds should be lower than that of a root disk,
and the specific volume's alert history is interesting from an operational
perspective. This also allows linking to the pertinent docs/dashboard
in the alert.

Bug: T261335
Change-Id: Ib74a7147330f1a6bea02397c47e7094ac9a0e29c