shared-storage: add specific NFS volume monitoring for cleanups
While we could change the parameters around the disk check for the servers,
that doesn't correctly recognize that we need different monitors for the
shared volumes. The thresholds should be lower than that of a root disk,
and the specific volume's alert history is interesting from an operational
perspective. This also allows linking to the pertinent docs/dashboard
in the alert.

This version fixes the spaces that broke change Ib74a7147330f1a6bea0

Bug: T261335
Change-Id: I52163f2a7e292e1df82873f65df884b064d46adc