wmcs: expose metricsinfra alert manager
Add configuration for the Prometheus vhost running in the metricsinfra
project that will reverse proxy a subset of the HTTP endpoints exposed
by the alertmanager component. Specifically the config blocks access to
service management, debugging, and state changing actions in the
alertmanager api. This config will allow a karma
(<https://github.com/prymitive/karma>) instance to display alerts and
silences that are active, but it will not allow karma to be used to
create or edit silences.

Change-Id: Ie44feef9a956699a9b913a68043e8f8206145d62