dynamicproxy: un-nest /favicon.ico and /robots.txt locations
Follow up to 8813762.

* Hoist settings for reverse proxy usage, including the default value of
  $backend, from the "location /" block to its parent "server"
  scope.
* Hoist location blocks for /favicon.ico and /robots.txt to their
  grandparent "server" scope.
* Use `access_by_lua_file` in each reverse proxy scope to update
  $backend if needed before calling `proxy_pass $backend;`

Nested locations worked in limited testing, but those tests did not
recreate the entire urlproxy config. In the live config, the $backend
variable is resolving to "" in the `proxy_redirect $backend;`
statements from the nested location blocks. After more in-depth testing,
even if $backend is defined directly in an inner location block the
proxy_redirect command is not provided the current value of that
variable (or possibly the definition of the variable is silently
ignored). This ends up looking a lot like
<https://trac.nginx.org/nginx/ticket/1383#comment:1> which says in part:
"This behaviour is basically identical to a nested location block."

The fix taken here is to place as much common config as possible into
the "server" scope and then un-nest the locations. Inheritance from
"server" to a child "location" seems to work as expected in testing and
the various blog posts and StackOverflow answers that could be found for
related issues. The new top level locations for /favicon.ico and
/robots.txt are also using the "=" modifier so that they only apply to
those exact paths.

Bug: T251628
Change-Id: I2aeaf754cd857c8f2857c159280067fabf9ff337