mediawiki::php::restarts: Allow disabling of php-fpm restarts
What:
Setting profile::mediawiki::php::restarts::opcache_limit to 0,
disables any opcache restarts as well as apcu clearing,

Why:

If a server had php-fpm restarts enabled at one point, we can't
have them properly disabled even if we set

profile::mediawiki::php::restarts::ensure: absent

The problems are:

a) the systemd timer is not removed but is rather left
in a failed state

b) scap tries to  "sudo php-check-and-restart"
during deployment, even though the script is not there.
While working on T253673, we might want to have the anti-corruption
measures temporarily disabled on some servers.

This is not the best solution, but it should be ok for now.

Bug: T261167
Bug: T253673
Change-Id: I7d277b96a7c798f98a65faf3bb40bab7bd3e69b4