cloud-vps: Add python3 client packages in cloud
This was held up because it didn't work on jessie. At this point,
jessie is gone on hardware and puppet is broken on jessie in VMs
regardless, so they don't entirely matter? Including keystoneauth1
because using keystoneclient without it is typically deprecated, and
I'm tired of manually installing it.

I *think* this should not be disasterous at this point.

Bug: T218423
Bug: T218426
Change-Id: I6cb62fd853f3719602d08bd8b2432497203959a9