phabricator: remove hiera() lookup from module
The goal here is to not have any hiera() inside
modules and instead only in profile parameters.

Also, to replace hiera() with lookup().

First I was about to add this as a parameter in the profile
class and pass it through but I also noticed there is no
"phabricator_servername" in Hiera.

We do have "phabricator_server" in Hiera though.

Is this the better and easier fix or should we use
phabricator_server from Hiera and not get it from
the config data... or both.

Change-Id: Ie19e90b816500afeffeed943d3ed4f72023f957c