wmcs-ceph-migrate: add a resize step
In order to properly apply ceph-related throttles we
'resize' the VM to a ceph-defined flavor after  migration.
The resize doesn't actually change size but it does change
the flavor and rebuild with desired throttles.

Note that most of the target flavors listed in the flavor
map here don't yet actually exist.

This patch also updates to python3, which didn't involve any
actual coding changes.

Bug: T261252
Change-Id: If578cece83160c44c0297d45c7b20eef17aa1d64