service.yaml: add releases as a service without LVS
This is an attempt to make releases.wikimedia.org
active-active after I was pointed to I876a03497f2e6a.

There are existing VMs, one per DC and they only host
static files.

The files are automatically rsynced between them (with --delete
to be actual mirrors).

There is a definition of "active server" in Hiera but that
is only for the purpose of deciding where users should
upload to and is considered the rsync source.

There already is a releases.discovery.wmnet DNS record
and only that is used in ATS config.

But, like other misc services without LVS and since we switched
from Varnish to ATS, that discovery record is just a CNAME
to one of 2 instances and would have to be switched manually.

This config stanza has been copied from chart-museum above.

The releases setup seems very similar to chart-museum with
a webserver and 2 Ganeti VMs and static files.

The assumption is that this goes with a DNS change like
I91b0871ffc54d5e228.

Change-Id: I18b06a61e352914b003b454da5f9d0244c1bc679