confd: only read from the master during the switchover
Etcd2 (codfw) is much slower than etcd3 (eqiad) in writing. So when
replicating from eqiad to codfw, and also changing a lot of records,
there can be significant lag between when things get written in eqiad,
and when that happens in codfw. This isn't a problem normally, but
it's an issue when switching over things from one datacenter to the
other as quickly as possible.

So temporarily switch all *confd* reads to eqiad.

Bug: T260889
Change-Id: Ie1f882c01a9d0df57e51163b3ed9c25bff7ca7e6