mariadb-backups: Update backup logic to check errors on log
Despite being coded, the logic for checking error messages on logs
was unused. This allowed errors on es4 and es5 to go on, as the
metadata files was anyway generated correctly. Fix this gap
and make sure dumps logs are checked before considering the
backup successful.

Also alter slightly the signature of the method.

This is the production hot fix for the patch at:
https://gerrit.wikimedia.org/r/c/operations/software/wmfbackups/+/623532

Bug: T138562
Change-Id: I271c2f9176c595651c06536ff81d18b934a94701