nftables: use wmflib::ensure for the service parameter
This is because systemd::service calls internally ::service which then undoes the masking operation
we are trying to do in the ::nftables class. 

Having the running/stopped paramenter makes things more complex. This patch changes the datatype to
use wmflib::ensure isntead, which allow us to reuse the parameter for systemd::service.

With the 'absent' parameter passed down to systemd::service and then ::service, this race should
not happen anymore.

Bug: T261724
Change-Id: Ib0a07fd3ae05e3f94db10f82bed63845e3b40cb1
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>