role:mx: add script to generate otrs aliases
This script runs every hour and creates a flat file containing email
addresses handled by OTRS.

In the current system when an email arrives we check it against OIT
ldap and if we find a match we deliver the mail to the gmail smtp
servers.  if there is no match then the address is passed to the next
router.  When it hits the otrs router we run a sql query to see if
the address exists in the otrs database, if it does send it to OTRS

The change I'm suggesting moves the gmail router to the last entry
so that if an email is not matched we just blindly send it to the
gmail servers and let them handle it.  This is mostly fine but it
would allow an OTRS admin to essentially steal an email address for
instance by adding an otrs alias for noc@wikimedia.org.

The aim of this script it to prevent that by pre-generating the list
of OTRS aliases, rejecting any values received from the database which
are also valid email addresses on the gmail servers.  Currently the
script finds no such entries as such all emails currently configured in
otrs will remain valid*.

* it is worth mentioning (although unrelated to this work) that the otrs
databases has entries for email addresses in domains which will never
hit our mail servers. e.g. wikimedia.de

Hosts: O:mail::mx
Bug: T244792
Change-Id: I3becf2218dd923cc0234ce2a4a5a2150f3acaf3b