wikidough: add an option to set the landing page
Wikidough accepts queries on /dns-query, and / currently returns a 404
as the path is not defined. This commit adds an option to accept queries
on /, to return a response as specified by newDOHResponseMapEntry() in
dnsdist.

The current text points to the Wikitech page for Wikidough; in the
future, this will be replaced by a privacy policy and/or a manual for
the project.

Bug: T252132
Change-Id: Ic1c275807cde9d2a2520da779fed837b45d399f3