role::kafka::jumbo::broker: move down inclusion of profile::standard
While reimaging to Buster I noticed that the interface::add_ip6_mapped
class (included in profile::standard) causes a stall in the first puppet
run, mostly when ferm rules are also added (see T255123#6425932).

This change moves profile::standard as last inclusion of the role, as
attempt to see if the puppet stall is a race condition of some sort.
Also removing the labs guard since it shouldn't be needed anymore.

Bug: T255123
Change-Id: I6503bb394a39ef99383dceec198804cd0aa4ba56