Add apache config for api.wikimedia.org
api.wikimedia.org is the new wiki that will act as documentation for
the API Gateway. it has already been configured on MediaWiki, so now
we want to make it accessible to the public.

This domain is currently accessed via Envoy running in k8s rather than
via the traditional path, but we still need appservers to serve it as
normal.

Previous deploy failed because we did not explicitly enable the site
in the wikimedia.conf file.

Bug: T246945
Change-Id: I95f6f6ea272eb77c6758aa3bdf3c53c45796603c