install_server: add reuse recipe for stat100x hosts with 4 disks
This recipe should allow the Analytics team to move stat100[4,6,7] hosts
to Debian Buster, preserving the /srv partition intact (where all the
home directories of users are).
After a chat with Kormat we agreed that this is not really "officially"
supported by the reuse script, since there is no guarantee that the kernel
will use the same mdX label for a set of disks in raid.
The plan is to backup the /srv partition of the host to reimage on another stat100x
node first, then attempt the upgrade and see.

Bug: T255028
Change-Id: Ia566710ecc14754546e62907ec5a08019168f3e5