dumps: rename the do_acme parameter and lookup
The do_acme parameter was used by the
letsencrypt::cert::integrated class but that is not being
used anymore. Instead acme_chief is used which does not
have this option.

So this has no relation amymore to actually doing a cert
request on a backend host and is kind of confusingly named
now.

This should now be a noop and just a rename of the do_acme parameter.

Change-Id: I587ce18a30d3a7760c62ac05ffdd3445d0aa3195