deployment-prep: add profile::prometheus::memcached_exporter::arguments
This is a reply to T244776#6364483. It should fix puppet runs on
cloud instances in this project using memached_exporter.

Unfortunately lookups that work for prod don't work for cloud :( (-> T255787)

After this there should be no more need to manually add this in Horizon web UI
and since it is in common.yaml the host name of the instance will not matter
in the future.

Bug: T244776
Change-Id: I21d48f96dadd386e433946d517aa6dc83dd7532b