webperf: add parameter to disable timing beacon monitoring in eqiad
There are different ways to do this and they all have their own issues.

For example adding a default value is slightly against style guide
but without it you will have to add the Hiera key explictly in
the context of both DCs. common/eqiad and common/codfw and without
it the class won't work. Let me know which you prefer.

The method to actually read the active DC state from conftool
and make this fully automatic is so far only used by mwmaint
servers and puppetmasters and has some comments that it is
"kind of an antipattern" as well.

A third way that many misc services do it is to put an "active_server"
in hiera common.yaml and then compare in puppet code to $::fqdn and
base decisions on that.

Eventually I decided against that and making the parameter explicit
to this one Icinga check we actually care about.

All other alerts seem to be fine in both DCs.
Of course it could also be something like "Boolean $has_monitoring".

Yea, so this for now even though it's not perfect.

Goal is to remove the Icinga alert in eqiad and this is how it
will work and we have done it for many other checks before.

Change-Id: Iaef79cf33d85d3cd2684d45c08cfdc46661a58f2